refactor(protocol-designer): Rearrange Step form fields and clarify
ahead of large functional changes in the push towards the new step form, this branch does some
housekeeping in the StepEditForm directory and a bit of styling.
